requirements.txt
flask==3.0.0
supabase==2.3.4
firebase-admin==6.4.0
python-dotenv==1.1.1
yfinance==0.2.28
psutil
requests==2.31.0
tweepy==4.16.0
newsapi-python==0.2.7
textblob==0.19.0
nltk==3.9.1
pandas>=1.3.0
numpy>=1.16.5
plotly==6.2.0
beautifulsoup4>=4.11.1
lxml>=4.9.1
gunicorn==21.2.0
google-generativeai
PyPDF2

# AI/PDF Analysis Dependencies
google-generativeai>=0.3.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Pillow>=9.0.0

# Additional PDF Processing
pdf2image>=1.16.0